[!!!][TASK] Move events to dedicated namespaces
