+++ src/Event/Crawler/CrawlingStarted.php
@@ -21,24 +21,37 @@ declare(strict_types=1);
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
-namespace EliasHaeussler\CacheWarmup\Event;
+namespace EliasHaeussler\CacheWarmup\Event\Crawler;
 
-use EliasHaeussler\CacheWarmup\Xml;
+use EliasHaeussler\CacheWarmup\Crawler;
+use EliasHaeussler\CacheWarmup\Sitemap;
 
 /**
- * ParserConstructed.
+ * CrawlingStarted.
  *
  * @author Elias Häußler <elias@haeussler.dev>
  * @license GPL-3.0-or-later
  */
-final readonly class ParserConstructed
+final readonly class CrawlingStarted
 {
+    /**
+     * @param list<Sitemap\Url> $urls
+     */
     public function __construct(
-        private Xml\Parser $parser,
+        private array $urls,
+        private Crawler\Crawler $crawler,
     ) {}
 
-    public function parser(): Xml\Parser
+    /**
+     * @return list<Sitemap\Url>
+     */
+    public function urls(): array
     {
-        return $this->parser;
+        return $this->urls;
+    }
+
+    public function crawler(): Crawler\Crawler
+    {
+        return $this->crawler;
     }
 }
