+++ src/Event/Parser/SitemapParsingFailed.php
@@ -21,44 +21,31 @@ declare(strict_types=1);
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
-namespace EliasHaeussler\CacheWarmup\Event;
+namespace EliasHaeussler\CacheWarmup\Event\Parser;
 
-use EliasHaeussler\CacheWarmup\Crawler;
-use EliasHaeussler\CacheWarmup\Result;
 use EliasHaeussler\CacheWarmup\Sitemap;
+use Throwable;
 
 /**
- * CrawlingFinished.
+ * SitemapParsingFailed.
  *
  * @author Elias Häußler <elias@haeussler.dev>
  * @license GPL-3.0-or-later
  */
-final readonly class CrawlingFinished
+final readonly class SitemapParsingFailed
 {
-    /**
-     * @param list<Sitemap\Url> $urls
-     */
     public function __construct(
-        private array $urls,
-        private Crawler\Crawler $crawler,
-        private Result\CacheWarmupResult $result,
+        private Sitemap\Sitemap $sitemap,
+        private Throwable $exception,
     ) {}
 
-    /**
-     * @return list<Sitemap\Url>
-     */
-    public function urls(): array
+    public function sitemap(): Sitemap\Sitemap
     {
-        return $this->urls;
+        return $this->sitemap;
     }
 
-    public function crawler(): Crawler\Crawler
+    public function exception(): Throwable
     {
-        return $this->crawler;
-    }
-
-    public function result(): Result\CacheWarmupResult
-    {
-        return $this->result;
+        return $this->exception;
     }
 }
